Quiz  App
Quizzery App: It is a Javascript based app in which I have created a quiz using different Javascript utilities, using event listeners and logic to generate and verify the inputs of the questions that are presented in the form of an array of objects. On verifying, it also renders the result for that specific user.

Tech Stack-HTML, CSS, Javascript